Refactor App.tsx to use MainSequencer component
